Render clipping before octree (#22)
Why: clipping is rendered over outlines and makes outlines less visible. Also toon outlines look pretty good on top of clipping planes.
One issue is that it doesn't work as expected when terrain is rendered as background - terrain is tinted, but not regular geometry.
For this we're planning to turn off terrain-as-background when there are clipping planes.