Option to disable wheel events when shift is down (#23)
It was earlier done with `enableShiftModifierOnWheel` inside flight control - instead of disabling the event - it zeroed the delta when calculation camera movement. The drawback is that it marked camera as moving which could mess with the web app.
Instead now we have `disableWheelOnShift` on input, so the event is just ignored completely when shift is pressed.